fix CI after adding repository
* moving module_config.yaml to the coin directory.
* update qtbase dependency.
* call qt_build_tests() in tests/CMakeLists.txt
* fix build on windows (correct type for QHasher, export Batch)
* put test in qt namespace (at least the QJsonValue operators are
  required to be in, but put everything)

+++ src/jsonrpc/qjsonrpcprotocol_p_p.h
@@ -67,7 +67,7 @@ template<typename T>
 struct QHasher
 {
     using argument_type = T;
-    using result_type = uint;
+    using result_type = size_t;
     result_type operator()(const argument_type &value) const { return qHash(value); }
 };
